move new builds monitor to regular background tasks
